+++ frontend/lost/src/components/SIA/lost-sia/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Prevent user from moving image out of canvas
 - Confirm label in LabelInput by click on the respective label
 - Configure name of the default Label by a prop 
+- Provide method to reset canvas zoom
 
 ### Fixed
 - Fixed jumping camera when zooming into the image